% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tv_aggregate_and_calculate_recall_results}
\alias{tv_aggregate_and_calculate_recall_results}
\title{Aggregate data and calculate recall results based on administrative level}
\usage{
tv_aggregate_and_calculate_recall_results(data, adm_level = "polling_station")
}
\arguments{
\item{data}{A data frame containing recall results}

\item{adm_level}{Character. Administrative level for aggregation and recall calculation}
}
\value{
Data frame with is_recalled column calculated at specified administrative level
}
\description{
This function aggregates recall data to the specified administrative level
and calculates is_recalled based on that level
}
\keyword{internal}
