% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tv_aggregate_and_calculate_winners}
\alias{tv_aggregate_and_calculate_winners}
\title{Aggregate data and calculate winners based on administrative level}
\usage{
tv_aggregate_and_calculate_winners(data, adm_level = "polling_station")
}
\arguments{
\item{data}{A data frame containing election results}

\item{adm_level}{Character. Administrative level for aggregation and winner calculation}
}
\value{
Data frame with is_elected column calculated at specified administrative level
}
\description{
This function aggregates election data to the specified administrative level
and calculates is_elected based on that level
}
\keyword{internal}
