% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source.R
\name{tv_cache_clear}
\alias{tv_cache_clear}
\title{Clear cached data}
\usage{
tv_cache_clear(file_name = NULL)
}
\arguments{
\item{file_name}{Character. Specific file name to remove. If NULL, removes all cached files.}
}
\description{
Remove cached data files. If no file_name is specified, removes all cached files.
}
