% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tv_calculate_election_winners}
\alias{tv_calculate_election_winners}
\title{Calculate election winners based on administrative level}
\usage{
tv_calculate_election_winners(data)
}
\arguments{
\item{data}{A data frame containing election results}
}
\value{
Data frame with is_elected column added
}
\description{
This function calculates is_elected based on the administrative scale
represented in the data. Winners are determined by highest vote count
at each unique combination of administrative units present in the data.
}
\keyword{internal}
