% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all.R
\name{tv_get_recall_2025_all}
\alias{tv_get_recall_2025_all}
\title{Get all 2025 legislator recall results}
\usage{
tv_get_recall_2025_all(level = c("village", "town"))
}
\arguments{
\item{level}{"village" (default) or "town" (aggregated).}
}
\value{
A tibble with all recall data at the specified level.
}
\description{
Get all 2025 legislator recall results
}
