% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{tv_list_available_elections}
\alias{tv_list_available_elections}
\title{List available Taiwan elections}
\usage{
tv_list_available_elections()
}
\value{
A tibble with information about available elections including:
\describe{
\item{year}{Election year}
\item{office}{Office type}
\item{sub_type}{Sub-type of election}
\item{description}{Description of the election}
}
}
\description{
Lists all available elections in the TaiwanVote package.
}
\examples{
\dontrun{
# List all available elections
tv_list_available_elections()
}

}
