% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{tv_list_available_recalls}
\alias{tv_list_available_recalls}
\title{List Available Recall Elections}
\usage{
tv_list_available_recalls()
}
\value{
A tibble with information about available recall elections including:
\describe{
\item{year}{Election year}
\item{office}{Office type}
\item{sub_type}{Sub-type of election}
\item{description}{Description of the recall election}
}
}
\description{
Lists all available recall elections in the TaiwanVote package.
}
\examples{
\dontrun{
# List all available recall elections
tv_list_available_recalls()
}

}
